Refactor: 유틸함수 분리 #17
